delete post

+++ controllers/appcontroller.js
@@ -352,10 +352,16 @@ const app_posts_post=async(req,res)=>{
 };
 //delete selected post by its _id
 const app_deletepost_post=async(req,res)=>{
+  
   var id =req.body._id;     // send the unique id made by mongo '_id' here it will find by that unique id and delete it
+  if(req.body.uploader==req.session.user.id)
+  {
   await Post.findByIdAndDelete(id)
   .then( (result)=>{res.send('deleted post')})
   .catch((err)=>{res.send(err)})
+  }else{
+    res.send("not your post")
+  }
 }
 //home data
 const app_home_get=async(req,res)=>{
